+++ src/types.ts
@@ -1,5 +1,5 @@
 import { JSONSchema } from 'mongochangestream'
-import type { ChangeStreamDocument, Document, ObjectId } from 'mongodb'
+import type { ChangeStreamDocument, ObjectId } from 'mongodb'
 import { Node } from 'obj-walker'
 
 interface RenameOption {
@@ -11,8 +11,15 @@ export interface ImmutableOption {
   /**
    * If the collection is immutable set this to true. This allows batch processing
    * where all change stream events are assumed to be inserts.
+   * @deprecated Use the autoOptimizeInserts option instead.
    */
   immutable?: boolean
+  /**
+   * Automatically optimize inserts by batching them together and flushing
+   * the insert queue when a non-insert event is received or a queue threshold
+   * is met - length, size in bytes, timeout, etc.
+   */
+  autoOptimizeInserts?: boolean
 }
 
 export interface SyncOptions extends RenameOption {